+++ packages/model/transaction_status.go
@@ -2,8 +2,8 @@ package model
 
 type TransactionStatus struct {
 	Hash      []byte `gorm:"primary_key;not null"`
-	Time      int32  `gorm:"not null;"`
-	Type      int32  `gorm:"not null"`
+	Time      int64  `gorm:"not null;"`
+	Type      int64  `gorm:"not null"`
 	WalletID  int64  `gorm:"not null"`
 	CitizenID int64  `gorm:"not null"`
 	BlockID   int64  `gorm:"not null"`
